FindMyWay/requirements.txt
streamlit>=1.0.0
Pillow>=8.0.0
simpleai>=0.7.0
matplotlib>=3.4.0
streamlit-drawable-canvas
